Implement filesystem as `intptr_t`

+++ src/xrCore/FS.h
@@ -139,14 +139,14 @@ public:
 	virtual CVirtualFileReader* cast_virtual_file_reader() { return nullptr; }
 	virtual IReader* cast_reader() { return nullptr; }
 
-	virtual u32			elapsed()	const = 0;
-	IC BOOL			eof			()	const		{return elapsed() <= 0;	};
+	virtual intptr_t		elapsed()	const = 0;
+	IC		BOOL			eof		()	const		{return elapsed() <= 0;	};
 
-	virtual u32		tell() const = 0;
-	virtual u32		length() const = 0;
-	virtual void	advance(int cnt) = 0;
+	virtual intptr_t		tell() const = 0;
+	virtual intptr_t		length() const = 0;
+	virtual void			advance(intptr_t cnt) = 0;
 
-	virtual void	r			(void* p, u32 cnt) = 0;
+	virtual void	r			(void* p, intptr_t cnt) = 0;
 
 	IC Fvector		r_vec3		()			{Fvector tmp;r(&tmp,3*sizeof(float));return tmp;	};
 	IC Fvector4		r_vec4		()			{Fvector4 tmp;r(&tmp,4*sizeof(float));return tmp;	};
@@ -192,11 +192,11 @@ public:
 		A.mul		(s);
 	}
 
-	virtual void	seek(int ptr) = 0;
+	virtual void	seek(intptr_t ptr) = 0;
 	// Set file pointer to start of chunk data (0 for root chunk)
 	IC	void		rewind		()			{	seek(0); }
 
-	virtual u32 find_chunk  (u32 ID, BOOL* bCompressed = 0);
+	virtual intptr_t find_chunk  (u32 ID, BOOL* bCompressed = 0);
 	
 	IC	BOOL		r_chunk		(u32 ID, void *dest)	// чтение XR Chunk'ов (4b-ID,4b-size,??b-data)
 	{
@@ -220,23 +220,24 @@ public:
 	u32 m_last_pos;
 };
 
-class XRCORE_API IReader :  public IReaderBase
+class XRCORE_API IReader :
+	public IReaderBase
 {
 protected:
 	char *			data	;
-	int				Pos		;
-	size_t			Size	;
-	int				iterpos	;
+	intptr_t		Pos		;
+	intptr_t		Size	;
+	intptr_t		iterpos	;
 
 public:
-	IC				IReader			()
+	IC IReader()
 	{
-		Pos			= 0;
+		Pos = 0;
 	}
 
-	virtual			~IReader		() {}
+	virtual ~IReader() {}
 
-	IC				IReader			(void *_data, int _size, int _iterpos=0)
+	IC IReader(void *_data, size_t _size, intptr_t _iterpos=0)
 	{
 		data		= (char *)_data	;
 		Size		= _size			;
@@ -256,15 +257,15 @@ protected:
 	u32 			advance_term_string			();
 
 public:
-	IC u32			elapsed		()	const override {	return Size-Pos;		};
-	   void			seek		(int ptr) override {	Pos=ptr; VERIFY((Pos<=Size) && (Pos>=0));};
-		u32			tell		()	const override {	return (u32)Pos;				};
-		u32			length		()	const override {	return (u32)Size;			};
-		void		advance		(int cnt) override {	Pos+=cnt;VERIFY((Pos<=Size) && (Pos>=0));};
+	IC intptr_t		elapsed		()	const override {	return Size - Pos;		};
+	   void			seek		(intptr_t ptr) override {	Pos=ptr; VERIFY((Pos<=Size) && (Pos>=0));};
+	   intptr_t		tell		()	const override {	return Pos;				};
+	   intptr_t		length		()	const override {	return Size;			};
+		void		advance		(intptr_t cnt) override {	Pos+=cnt;VERIFY((Pos<=Size) && (Pos>=0));};
 	IC void*		pointer		()	const		{	return &(data[Pos]);	};
 
 public:
-	void			r			(void *p,u32 cnt) override;
+	void			r			(void *p, intptr_t cnt) override;
 
 	void			r_string	(char *dest, u32 tgt_sz);
 	void			r_string	(xr_string& dest);
@@ -325,8 +326,7 @@ public:
 
 	// iterators
 	IReader*		open_chunk_iterator		(u32& ID, IReader* previous=NULL);	// NULL=first
-
-	u32 			find_chunk	(u32 ID, BOOL* bCompressed = 0) override;
+	intptr_t 		find_chunk	(u32 ID, BOOL* bCompressed = 0) override;
 
 private:
 	typedef IReaderBase inherited;
